+++ apps/docs/design/blocks/stack/index.tsx
@@ -1,15 +1,6 @@
-export default function Stack({
-  children,
-  gap,
-}: {
-  children: React.ReactNode
-  gap: any
-}) {
+export default function Stack({ children, gap }: { children: React.ReactNode; gap: any }) {
   return (
-    <div
-      className="stack"
-      style={{ display: 'flex', flexDirection: 'column', gap }}
-    >
+    <div className="stack" style={{ display: 'flex', flexDirection: 'column', gap }}>
       {children}
     </div>
   )