+++ libs/core/src/components/input/input.ts
@@ -77,7 +77,7 @@ export class GdsInput extends GdsFormControlElement<string> {
    * The maximum number of characters allowed in the field.
    */
   @property({ type: Number })
-  maxlength = Number.MAX_SAFE_INTEGER
+  maxlength: number = Number.MAX_SAFE_INTEGER
 
   /**
    * The variant of the input field. The default variant displays a label, supporting text, and
@@ -93,7 +93,7 @@ export class GdsInput extends GdsFormControlElement<string> {
   variant: 'default' | 'floating-label' = 'default'
 
   @property({ type: String })
-  size: 'default' | 'small' = 'default'
+  size: 'large' | 'small' = 'large'
 
   @queryAsync('input')
   private elInputAsync!: Promise<HTMLInputElement>